requirements.txt
# Interfaz web
streamlit==1.32.0
streamlit-chat==0.1.1

# OpenAI
openai==1.82.0

# Base de datos
mysql-connector-python==8.3.0
SQLAlchemy==2.0.28
pymysql==1.1.0

# Procesamiento de datos y visualización
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
altair==5.2.0
matplotlib==3.8.3
seaborn==0.13.2
bokeh==3.4.0

# Variables de entorno
python-dotenv==1.0.0

# Modelos de IA y NLP
transformers==4.38.2
torch==2.2.1
huggingface-hub==0.21.4
langchain>=0.1.12
langchain-community>=0.0.28
langchain-openai>=0.0.8
langchain-experimental>=0.0.49

# Utilidades
python-dateutil==2.8.2 